Hero Changes
Changed how the battleChoices() method works. It now uses a helper method in the hero class to run the battleChoices() menu. This was done to have less repeating code and also to make it much easier to swap out the Keyboard.java class when that time comes

Also, added feilds for the supercall values in order to decrease ambiguity on the code's call.

Changed Visibility of defend() in hero.java to private as defend() is only used in the hero.java class and doesn't need to be accessed elsewhere.